+++ R/utils.R
@@ -3,10 +3,6 @@ msg_to_raw <- function(msg) {
   serialize(msg, NULL)
 }
 
-msg_to_integer <- function(msg) {
-  as.integer(msg_to_raw(msg))
-}
-
 
 #' Custom stop and message functions
 #'