+++ NEWS.md
@@ -1,5 +1,9 @@
 # pqcrypto (development version)
 
+* Internals: 
+  - Some changes to classes, new tests, lib compilation, ...
+  - Change data types from integer to raw.
+
 # pqcrypto 0.1.0
 
 * Initial public version.